Crate orbit_server

Re-exports

pub use load_balancer::LoadBalancer;

pub use load_balancer::LoadBalancerStats;

pub use load_balancer::LoadBalancingStrategy;

pub use load_balancer::NodeLoad;

pub use mesh::AddressableDirectory;

pub use mesh::ClusterManager;

pub use mesh::ClusterStats;

pub use mesh::DirectoryStats;

pub use server::OrbitServer;

pub use server::OrbitServerBuilder;

pub use server::OrbitServerConfig;

pub use server::ProtocolConfig;

pub use server::ProtocolStats;

pub use server::ServerStats;

Modules

directory

load_balancer

Load balancing and actor placement algorithms for the Orbit cluster

mesh

Server-side mesh management and cluster coordination

persistence

Provider-based persistence system for Orbit server

server

Main Orbit server implementation for hosting actors and managing the cluster